Added nodesbylocation to __init__.py and removed from types.py to fix a bug where it wasn't found.
